Updating types in all files, cleanup and fixing bugs

+++ includes/src_includes/interpolation.h
@@ -3,4 +3,4 @@
 #include "matrix.h"
 
 
-void interpolate_vector(std::vector<float> coarse_array, std::vector<float> fine_array);
+void interpolate_vector(std::vector<double> coarse_array, std::vector<double> fine_array);
